toString removal v2 & lil refactor
