--- server/src/main/java/org/jetbrains/bsp/bazel/server/sync/languages/java/Jdk.kt
@@ -1,16 +1,0 @@
-package org.jetbrains.bsp.bazel.server.sync.languages.java
-
-import org.jetbrains.bsp.bazel.commons.Format
-import java.net.URI
-
-data class Jdk(
-    val version: String,
-    val javaHome: URI?
-) {
-
-    override fun toString(): String {
-        return Format.`object`(
-            "Jdk", Format.entry("version", version), Format.entry("javaHome", javaHome)
-        )
-    }
-}
