+++ server/src/main/java/org/jetbrains/bsp/bazel/server/sync/model/Label.kt
@@ -1,9 +1,3 @@
 package org.jetbrains.bsp.bazel.server.sync.model
 
-
-data class Label(val value: String) {
-
-    override fun toString(): String {
-        return value
-    }
-}
+data class Label(val value: String)
